+++ src/ui/point.hpp
@@ -1,15 +1,16 @@
 #pragma once
 
+#include "int.hpp"
 
 namespace shar {
 
 struct Point {
-  int x;
-  int y;
+  usize x;
+  usize y;
 
   static Point origin() {
     return Point{0, 0};
   }
 };
 
-}
+} // namespace shar
